requirements.txt
# Core dependencies
numpy>=1.24.0
pandas>=2.0.0
scipy==1.11.4

# Geospatial
rasterio==1.3.9
geopandas==0.14.1
shapely>=2.0.0
pyproj>=3.5.0

# Visualization
matplotlib==3.8.2
folium==0.15.1
contextily>=1.3.0

# API & Data
requests==2.31.0
python-dotenv==1.0.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.23.0

# Image processing
scikit-image>=0.21.0

# Testing
pytest==7.4.3

# Utilities
tqdm==4.66.1